requirements.txt
# Core dependencies
python-dotenv>=1.0.0
streamlit>=1.28.0
langchain>=0.0.340
langchain-community>=0.0.10
transformers>=4.35.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
spacy>=3.7.2
rank-bm25>=0.2.2
reportlab>=4.0.4
plotly>=5.17.0
fastapi>=0.104.1
uvicorn>=0.24.0
joblib>=1.3.2
pytest>=7.4.3
pydantic-settings>=2.0.0

# Additional dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
httpx>=0.25.0

# Document processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Caching
redis>=5.0.0
diskcache>=5.6.0

# Logging and monitoring
loguru>=0.7.0
rich>=13.0.0

# Testing
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
pre-commit>=3.4.0
